Remove duplication caused by Document enum cases
By dynamically modelling resources and collections as enum cases, we
lose the ability to statically determine whether we expect to decode a
single resource or a collection, and so we have to dynamically
destructure the document in order to restore the state guarantee.

Removing the enum cases and statically calling `decodeResource` or
`decodeCollection` gives all the safety we need and removes duplication.