Stop pretending to mutate relationships while decoding
The code as written makes it appear that we're removing an object from
the "included" array whenever a match is found. This doesn't appear to
actually be the case: as we recurse the tree of relationships, each node
appears to be acting on its own copy of the relationships, making the
mutation unnecessary.